refactor(tests): move test_overfit_causal_lm to its own file
